tweak to build on linux

+++ src/clustering/administration/logs/log_writer.cc
@@ -429,7 +429,7 @@ bool fallback_log_writer_t::write(const log_message_t &msg, std::string *error_o
         }
 #else
         ssize_t write_res = ::write(filefd, console_formatted.data(), console_formatted.length());
-        if (write_res != console_formatted.length()) {
+        if (write_res != static_cast<ssize_t>(console_formatted.length())) {
             error_out->assign("cannot write to stdout/stderr: " + errno_string(get_errno()));
             return false;
         }
